Extract the transcript, attachment, and context-pack statics
ChatTranscriptBuilder now owns the pure transcript derivations (blocks(from:),
displayUserText, safeFilename, and the checkpoint transcript digest);
ChatAttachmentLoader and ContextPackLoader own the nonisolated file readers
with their budget rules. All are UI-free static namespaces per the house
pattern, so their rules stay unit-testable without an AppModel. Call sites
and tests use the new homes directly; AppModel keeps one displayUserText
forwarder so InspectorView is untouched until its own migration commit.

Behavior-identical refactor; project regenerated with xcodegen.
